backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
weasyprint==52.5
google-generativeai==0.1.0
python-dotenv==0.19.0
pymongo==3.12.0
pdfminer.six==20201018
python-docx==0.8.11